+++ app/src/main/java/pl/llp/aircasting/data/local/entity/sensor_thresholds.kt
@@ -1,6 +1,9 @@
-package pl.llp.aircasting.data.local.data_classes
+package pl.llp.aircasting.data.local.entity
 
-import androidx.room.*
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
 import pl.llp.aircasting.data.model.MeasurementStream
 
 @Entity(
@@ -28,22 +31,4 @@ data class SensorThresholdDBObject(
         measurementStream.thresholdHigh,
         measurementStream.thresholdVeryHigh
     )
-}
-
-@Dao
-interface SensorThresholdDao {
-    @Query("SELECT * FROM sensor_thresholds")
-    fun getAll() : List<SensorThresholdDBObject>
-
-    @Query("SELECT * FROM sensor_thresholds WHERE sensor_name in (:sensorNames)")
-    fun allBySensorNames(sensorNames: List<String>): List<SensorThresholdDBObject>
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    fun insert(sensorThresholdDBObject: SensorThresholdDBObject): Long
-
-    @Query("UPDATE sensor_thresholds SET threshold_very_low=:thresholdVeryLow, threshold_low=:thresholdLow, threshold_medium=:thresholdMedium, threshold_high=:thresholdHigh, threshold_very_high=:thresholdVeryHigh WHERE sensor_name=:sensorName")
-    fun update(sensorName: String, thresholdVeryLow: Int, thresholdLow: Int, thresholdMedium: Int, thresholdHigh: Int, thresholdVeryHigh: Int)
-
-    @Query("DELETE FROM sensor_thresholds")
-    fun deleteAll()
 }
