+++ app/src/main/java/pl/llp/aircasting/data/model/observers/FixedSessionsObserver.kt
@@ -1,7 +1,7 @@
 package pl.llp.aircasting.data.model.observers
 
 import androidx.lifecycle.LifecycleOwner
-import pl.llp.aircasting.data.local.data_classes.SessionWithStreamsDBObject
+import pl.llp.aircasting.data.local.entity.SessionWithStreamsDBObject
 import pl.llp.aircasting.data.model.Session
 import pl.llp.aircasting.ui.viewmodel.SessionsViewModel
 import pl.llp.aircasting.ui.view.screens.dashboard.SessionsViewMvc
