+++ app/src/main/java/pl/llp/aircasting/util/helpers/services/AveragingService.kt
@@ -1,8 +1,8 @@
 package pl.llp.aircasting.util.helpers.services
 
 import android.util.Log
-import pl.llp.aircasting.data.local.data_classes.MeasurementDBObject
-import pl.llp.aircasting.data.local.data_classes.SessionDBObject
+import pl.llp.aircasting.data.local.entity.MeasurementDBObject
+import pl.llp.aircasting.data.local.entity.SessionDBObject
 import pl.llp.aircasting.data.api.repository.MeasurementStreamsRepository
 import pl.llp.aircasting.data.api.repository.MeasurementsRepository
 import pl.llp.aircasting.data.api.repository.SessionsRepository
